Moving common classes for tests in common so that they can be used in all projects
